move error box to the top
