+++ frontend/fingertips-frontend/components/forms/SearchForm/index.tsx
@@ -1,16 +1,15 @@
 'use client';
 
-import { searchIndicator, SearchFormState } from './searchActions';
+import { SearchFormState } from './searchActions';
 import {
-  ErrorSummary,
   InsetText,
   Button,
   InputField,
   Paragraph,
   H3,
+  Link,
 } from 'govuk-react';
 import { spacing } from '@govuk-react/lib';
-import { useActionState } from 'react';
 import styled from 'styled-components';
 
 const StyledInputField = styled(InputField)(
@@ -22,28 +21,7 @@ export const SearchForm = ({
 }: {
   searchFormState: SearchFormState;
 }) => {
-  const [state, formAction] = useActionState(searchIndicator, searchFormState);
-
   return (
-    <form action={formAction}>
-      {state.message && (
-        <ErrorSummary
-          description="At least one of the following fields must be populated:"
-          errors={[
-            {
-              targetName: 'indicator',
-              text: 'Indicator field',
-            },
-          ]}
-          data-testid="search-form-error-summary"
-          onHandleErrorClick={(targetName: string) => {
-            const indicator = document.getElementById(targetName);
-            indicator?.scrollIntoView(true);
-            indicator?.focus();
-          }}
-        />
-      )}
-      <br />
       <div
         data-testid="search-form"
         style={{ backgroundColor: '#ddd', padding: '20px 20px 0px 20px' }}
@@ -60,7 +38,7 @@ export const SearchForm = ({
           input={{
             id: 'indicator',
             name: 'indicator',
-            defaultValue: state.indicator ?? searchFormState.indicator,
+            defaultValue: searchFormState.indicator,
           }}
           hint={
             <div style={{ color: 'black' }}>
@@ -68,50 +46,39 @@ export const SearchForm = ({
             </div>
           }
           meta={{
-            touched: !!state.message,
+            touched: !!searchFormState.message,
             error: 'This field value may be required',
           }}
           data-testid="search-form-input-indicator"
         >
           Search by subject
         </StyledInputField>
-        <>
-          <StyledInputField
-            style={{ marginBottom: '0' }}
-            input={{
-              id: 'indicator',
-              name: 'indicator',
-              defaultValue: state.indicator ?? searchFormState.indicator,
-            }}
-            hint={
-              <div style={{ color: 'black' }}>
-                For example postcode, county, local authority, NHS Trust or
-                General Practice name or code
-              </div>
-            }
-            meta={{
-              touched: !!state.message,
-              error: 'This field value may be required',
-            }}
-            data-testid="search-form-input-area"
-          >
-            Search for an area by location or organisation
-          </StyledInputField>
-          <p
-            style={{
-              textDecoration: 'underline',
-              margin: '8px 0px 0px',
-              fontSize: '18px',
-            }}
-          >
-            Or filter by area
-          </p>
-        </>
-        <br />
-        <Button type="submit" data-testid="search-form-button-submit">
+        <StyledInputField
+          style={{ marginBottom: '5px' }}
+          input={{
+            id: 'areaSearched',
+            name: 'areaSearched',
+            defaultValue: searchFormState.areaSearched,
+          }}
+          hint={
+            <div style={{ color: 'black' }}>
+              For example postcode, county, local authority, NHS Trust or
+              General Practice name or code
+            </div>
+          }
+          meta={{
+            touched: !!searchFormState.message,
+            error: 'This field value may be required',
+          }}
+          data-testid="search-form-input-area"
+        >
+          Search for an area by location or organisation
+        </StyledInputField>
+        <Link href='#' data-testid="search-form-link-filter-area">Or filter by area</Link>
+        <br/>
+        <Button type="submit" data-testid="search-form-button-submit" style={{marginTop: '25px'}}>
           Search
         </Button>
       </div>
-    </form>
   );
 };
