+++ frontend/fingertips-frontend/components/forms/SearchForm/searchActions.ts
@@ -9,11 +9,13 @@ const $SearchFormSchema = z.object({
     .string()
     .trim()
     .min(1, { message: 'Please enter an indicator id' }),
+  areaSearched: z.string().optional()
 });
 
 export type State = {
   errors?: {
     indicator?: string[];
+    areaSearched?: string[];
   };
   message?: string | null;
 };
@@ -28,11 +30,13 @@ export async function searchIndicator(
 ): Promise<SearchFormState> {
   const validatedFields = $SearchFormSchema.safeParse({
     indicator: formData.get('indicator'),
+    areaSearched: formData.get('areaSearched'),
   });
 
   if (!validatedFields.success) {
     return {
       indicator: formData.get('indicator')?.toString().trim() ?? '',
+      areaSearched: formData.get('areaSearched')?.toString().trim() ?? '',
       errors: validatedFields.error.flatten().fieldErrors,
       message: 'Please enter a value for the indicator field',
     };
