+++ frontend/fingertips-frontend/components/pages/home/home.test.tsx
@@ -1,9 +1,12 @@
-import { render, screen } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
 import { Home } from '@/components/pages/home/index';
 import { SearchFormState } from '@/components/forms/SearchForm/searchActions';
+import { expect } from '@jest/globals';
+import { userEvent } from '@testing-library/user-event';
 
 const initialState: SearchFormState = {
   indicator: '',
+  areaSearched: '',
   message: null,
   errors: {},
 };
@@ -33,4 +36,46 @@ it('should render the SearchForm component', () => {
   const searchForm = screen.getByTestId('search-form');
 
   expect(searchForm).toBeInTheDocument();
+});
+
+it('should display the error summary component when there is a validation error', () => {
+  const errorState: SearchFormState = {
+    indicator: '',
+    areaSearched: '',
+    message: 'Error message',
+    errors: {},
+  };
+
+  render(<Home searchFormState={errorState} />);
+
+  expect(screen.getByTestId('search-form-error-summary')).toBeInTheDocument();
+});
+
+it('should display the error summary component when there is a validation error', async () => {
+  // Add missing function to jsdom
+  const scrollMock = jest.fn();
+  window.HTMLElement.prototype.scrollIntoView = scrollMock;
+
+  const user = userEvent.setup();
+
+  const errorState: SearchFormState = {
+    indicator: '',
+    areaSearched: '',
+    message: 'Error message',
+    errors: {},
+  };
+
+  render(<Home searchFormState={errorState} />);
+  
+  
+
+  const anchor = screen.getByText('Indicator field').closest('a');
+  if (anchor) {
+    await user.click(anchor);
+  }
+
+  await waitFor(() => {
+    expect(screen.getByRole('textbox', { name: /indicator/i })).toHaveFocus();
+  });
+  expect(scrollMock).toBeCalledTimes(1);
 });
